refactor(debugger): share storage access helper
